Rename Canonical to ContainsEntity.
