+++ src/registry/mod.rs
@@ -24,7 +24,6 @@
 //! [`registry!`]: crate::registry!
 //! [`World`]: crate::world::World
 
-mod canonical;
 mod contains;
 mod debug;
 mod eq;
@@ -36,8 +35,7 @@ mod sync;
 
 #[cfg(feature = "serde")]
 pub(crate) use self::serde::{RegistryDeserialize, RegistrySerialize};
-pub(crate) use canonical::Canonical;
-pub(crate) use contains::ContainsEntities;
+pub(crate) use contains::{ContainsEntities, ContainsEntity};
 pub(crate) use debug::RegistryDebug;
 pub(crate) use eq::{RegistryEq, RegistryPartialEq};
 pub(crate) use send::RegistrySend;
